+++ AppiumLibrary/keywords/_touch.py
@@ -1,9 +1,15 @@
 # -*- coding: utf-8 -*-
 
-from appium.webdriver.common.touch_action import TouchAction
+import time
 
+from appium.webdriver.extensions.action_helpers import ActionHelpers
+
+from datetime import timedelta
 from AppiumLibrary.locators import ElementFinder
 from .keywordgroup import KeywordGroup
+
+from robot.api import logger
+from typing import Union
 
 
 class _TouchKeywords(KeywordGroup):
@@ -20,39 +26,45 @@ class _TouchKeywords(KeywordGroup):
         element = self._element_find(locator, True, True)
         driver.zoom(element=element, percent=percent, steps=steps)
 
-    def pinch(self, locator, percent="200%", steps=1):
-        """*DEPRECATED!!* use `Execute Script` instead.
-        Pinch in on an element a certain amount.
-        """
-        driver = self._current_application()
-        element = self._element_find(locator, True, True)
-        driver.pinch(element=element, percent=percent, steps=steps)
-
-    def swipe(self, start_x, start_y, offset_x, offset_y, duration=1000):
+    def swipe(self, start_x: Union[int, float], start_y: Union[int, float], end_x: Union[int, float], end_y: Union[int, float], duration: Union[int, timedelta] = timedelta(seconds=1)):
         """
         Swipe from one point to another point, for an optional duration.
 
         Args:
-         - start_x - x-coordinate at which to start
-         - start_y - y-coordinate at which to start
-         - offset_x - x-coordinate distance from start_x at which to stop
-         - offset_y - y-coordinate distance from start_y at which to stop
-         - duration - (optional) time to take the swipe, in ms.
+        - start_x: x-coordinate at which to start
+        - start_y: y-coordinate at which to start
+        - end_x: x-coordinate at which to stop
+        - end_y: y-coordinate at which to stop
+        - duration: defines the swipe speed as time taken to swipe from point a to point b.
 
-        Usage:
-        | Swipe | 500 | 100 | 100 | 0 | 1000 |
 
-        _*NOTE: *_
-        Android 'Swipe' is not working properly, use ``offset_x`` and ``offset_y`` as if these are destination points.
+        Examples:
+        | Swipe | 500 | 100 | 100 | 0 | 1s |
+        | Swipe | 500 | 100 | 100 | 0 | 100ms |
         """
-        x_start = int(start_x)
-        x_offset = int(offset_x)
-        y_start = int(start_y)
-        y_offset = int(offset_y)
-        driver = self._current_application()
-        driver.swipe(x_start, y_start, x_offset, y_offset, duration)
 
-    def swipe_by_percent(self, start_x, start_y, end_x, end_y, duration=1000):
+        if isinstance(duration, int):
+            logger.warn(
+                "Keyword 'Swipe' will not support int in ms for 'duration' in the future. "
+                "Use timedelta with units ('ms' or 's') instead."
+            )
+            duration = timedelta(milliseconds=duration)
+
+        args = [start_x, start_y, end_x, end_y]
+
+        for i, arg in enumerate(args):
+            if isinstance(arg, float):
+                logger.warn(
+                    "Keyword 'Swipe' converts the values of 'start_x', 'start_y', 'end_x', 'end_y' to integer."
+                )
+                args[i] = int(arg)
+
+        start_x, start_y, end_x, end_y = args
+
+        driver = self._current_application()
+        driver.swipe(start_x, start_y, end_x, end_y, duration.total_seconds() * 1000)
+
+    def swipe_by_percent(self, start_x: Union[int, float], start_y: Union[int, float], end_x: Union[int, float], end_y: Union[int, float], duration: Union[int, timedelta] = timedelta(seconds=1)):
         """
         Swipe from one percent of the screen to another percent, for an optional duration.
         Normal swipe fails to scale for different screen resolutions, this can be avoided using percent.
@@ -62,22 +74,40 @@ class _TouchKeywords(KeywordGroup):
          - start_y - y-percent at which to start
          - end_x - x-percent distance from start_x at which to stop
          - end_y - y-percent distance from start_y at which to stop
-         - duration - (optional) time to take the swipe, in ms.
+         - duration - (optional) time to take the swipe
 
-        Usage:
+        Examples:
         | Swipe By Percent | 90 | 50 | 10 | 50 | # Swipes screen from right to left. |
 
         _*NOTE: *_
         This also considers swipe acts different between iOS and Android.
 
-        New in AppiumLibrary 1.4.5
         """
+
+        if isinstance(duration, int):
+            logger.warn(
+                "Keyword 'Swipe By Percent' will not support int in ms for 'duration' in the future. "
+                "Use timedelta with units ('ms' or 's') instead."
+            )
+            duration = timedelta(milliseconds=duration)
+
+        args = [start_x, start_y, end_x, end_y]
+
+        for i, arg in enumerate(args):
+            if isinstance(arg, float):
+                logger.warn(
+                    "Keyword 'Swipe' converts the values of 'start_x', 'start_y', 'end_x', 'end_y' to integer."
+                )
+                args[i] = int(arg)
+
+        start_x, start_y, end_x, end_y = args
+
         width = self.get_window_width()
         height = self.get_window_height()
-        x_start = float(start_x) / 100 * width
-        x_end = float(end_x) / 100 * width
-        y_start = float(start_y) / 100 * height
-        y_end = float(end_y) / 100 * height
+        x_start = int(start_x / 100 * width)
+        x_end = int(end_x / 100 * width)
+        y_start = int(start_y / 100 * height)
+        y_end = int(end_y / 100 * height)
         x_offset = x_end - x_start
         y_offset = y_end - y_start
         platform = self._get_platform()
@@ -97,42 +127,115 @@ class _TouchKeywords(KeywordGroup):
         driver = self._current_application()
         driver.scroll(el1, el2)
 
-    def scroll_down(self, locator):
-        """Scrolls down to element"""
+    def scroll_down(self, locator, timeout=10, retry_interval=1):
+        """Scrolls down until the element is found or until the timeout (Android only) is reached.
+            Args:
+        - ``locator`` - (mandatory)  Locator of the element to scroll down to.
+        - ``timeout`` - (optional) timeout in seconds (default 10 seconds) - Android only
+        - ``retry_interval`` - (optional) interval between scroll attempts in seconds (default one second)
+        """
         driver = self._current_application()
-        element = self._element_find(locator, True, True)
-        driver.execute_script("mobile: scroll", {"direction": 'down', 'elementid': element.id})
+        platform = self._get_platform()
+        if platform == 'android':
+            start_time = time.time()
+            while time.time() - start_time < timeout:
+                try:
+                    element = self._element_find(locator, True, True)
+                    return True
+                except ValueError:
+                    print('Element not visible, scrolling...')
+                    width = self.get_window_width()
+                    height = self.get_window_height()
 
-    def scroll_up(self, locator):
-        """Scrolls up to element"""
+                    x = width / 2
+                    start_y = height * 0.8 # 80% of the screen
+                    end_y = height * 0.2 # 20% of the screen
+
+                    driver.swipe(int(x), int(start_y), int(x), int(end_y), 1000)
+                time.sleep(retry_interval)
+        else:
+            element = self._element_find(locator, True, True)
+            driver.execute_script("mobile: scroll", {"direction": 'down', 'elementid': element.id})
+            return True
+
+        raise AssertionError(f"Element '{locator}' not found within {timeout} seconds.")
+
+    def scroll_up(self, locator, timeout=10, retry_interval=1):
+        """Scrolls up until the element is found or the timeout (Android only) is reached.
+            Args:
+        - ``locator`` - (mandatory)  Locator of the element to scroll down to.
+        - ``timeout`` - (optional) timeout in seconds (default 10 seconds) - Android only
+        - ``retry_interval`` - (optional) interval between scroll attempts in seconds (default one second)
+        """
         driver = self._current_application()
-        element = self._element_find(locator, True, True)
-        driver.execute_script("mobile: scroll", {"direction": 'up', 'elementid': element.id})
+        platform = self._get_platform()
+        if platform == 'android':
+            start_time = time.time()
+            while time.time() - start_time < timeout:
+                try:
+                    element = self._element_find(locator, True, True)
+                    return True
+                except ValueError:
+                    print('Element not visible, scrolling...')
+                    width = self.get_window_width()
+                    height = self.get_window_height()
+
+                    x = width / 2
+                    start_y = height * 0.2
+                    end_y = height * 0.8
+
+                    driver.swipe(int(x), int(start_y), int(x), int(end_y), 1000)
+                time.sleep(retry_interval)
+        else:
+            element = self._element_find(locator, True, True)
+            driver.execute_script("mobile: scroll", {"direction": 'up', 'elementid': element.id})
+            return True
+
+        raise AssertionError(f"Element '{locator}' not found within {timeout} seconds.")
+
 
     def long_press(self, locator, duration=1000):
-        """*DEPRECATED!!* Since selenium v4, use other keywords.
+        """Long press the element identified by ``locator`` with optional ``duration``.
 
-        Long press the element with optional duration """
-        driver = self._current_application()
+        Args:
+        - ``locator`` - (mandatory)
+        - ``duration`` - duration of time to tap, in ms. Default: 1000ms
+
+        Examples:
+        | Long Press | xpath=//*[@resource-id='login_button'] |
+        | Long Press | xpath=//*[@name='link'] | duration=3000 |
+        """
         element = self._element_find(locator, True, True)
-        action = TouchAction(driver)
-        action.press(element).wait(duration).release().perform()
+        location = element.location
+        size = element.size
+        center_x = location['x'] + size['width'] // 2
+        center_y = location['y'] + size['height'] // 2
+        driver = self._current_application()
+        driver.tap([(center_x, center_y)], duration)
 
-    def tap(self, locator, x_offset=None, y_offset=None, count=1):
-        """*DEPRECATED!!* Since selenium v4, use `Tap With Positions` keyword.
-
+    def tap(self, locator, count=1, duration=500):
+        """
         Tap element identified by ``locator``.
 
         Args:
-        - ``locator`` - (mandatory). Taps coordinates when set to ${None}.
-        - ``x_offset`` - (optional) x coordinate to tap, relative to the top left corner of the element.
-        - ``y_offset`` - (optional) y coordinate. If y is used, x must also be set, and vice versa
+        - ``locator`` - (mandatory)
         - ``count`` - can be used for multiple times of tap on that element
+        - ``duration`` - duration of time to tap, in ms. Default: 500ms
+
+        Examples:
+        | Tap | xpath=//*[@resource-id='login_button'] |
+        | Tap | xpath=//*[@name='picture'] | duration=100
+        | Tap | xpath=//*[@name='picture'] | count=2 | duration=100
+
         """
         driver = self._current_application()
-        el = self._element_find(locator, True, True)
-        action = TouchAction(driver)
-        action.tap(el,x_offset,y_offset, count).perform()
+        for _ in range(count):
+            element = self._element_find(locator, True, True)
+            location = element.location
+            size = element.size
+            center_x = location['x'] + size['width'] // 2
+            center_y = location['y'] + size['height'] // 2
+            driver.tap([(center_x, center_y)], duration)
 
     def tap_with_positions(self, duration=500, *locations):
         """Taps on a particular place with up to five fingers, holding for a
@@ -149,15 +252,14 @@ class _TouchKeywords(KeywordGroup):
         |  @{fingerPositions}  |  create list  |  ${firstFinger}  |  ${secondFinger}  |
         |  Sleep  |  1  |
         |  Tap with Positions  |  ${1000}  |  @{fingerPositions}  |
-
-        New in AppiumLibrary v2
         """
         driver = self._current_application()
         driver.tap(positions=list(locations), duration=duration)
-        
+
     def tap_with_number_of_taps(self, locator, number_of_taps, number_of_touches):
-        """ Sends one or more taps with one or more touch points.iOS only.
-        
+        """ Sends one or more taps with one or more touch points\n
+        *iOS only.*
+
         Args:
         - ``number_of_taps`` - The number of taps.
         - ``number_of_touches`` - The number of touch points.
@@ -168,40 +270,18 @@ class _TouchKeywords(KeywordGroup):
         driver.execute_script("mobile: tapWithNumberOfTaps", params)
 
     def click_alert_button(self, button_name):
-        """ Clicks on Alert button identified by Name.iOS only.
+        """ Clicks on Alert button identified by Name.\n
+        *iOS only.*
 
         Args:
         - ``button_name`` - Text on the iOS alert button.
 
         Example:
         |  Click Alert Button  |  Allow  |
-
-        New in AppiumLibrary v2
         """
         driver = self._current_application()
         params={'action': 'accept', 'buttonLabel': button_name}
         driver.execute_script("mobile: alert", params)
-
-    def click_a_point(self, x=0, y=0, duration=100):
-        """*DEPRECATED!!* Since selenium v4, use other keywords.
-
-        Click on a point"""
-        self._info("Clicking on a point (%s,%s)." % (x,y))
-        driver = self._current_application()
-        action = TouchAction(driver)
-        try:
-            action.press(x=float(x), y=float(y)).wait(float(duration)).release().perform()
-        except:
-            assert False, "Can't click on a point at (%s,%s)" % (x,y)
-
-    def click_element_at_coordinates(self, coordinate_X, coordinate_Y):
-        """*DEPRECATED!!* Since selenium v4, use other keywords.
-
-        click element at a certain coordinate """
-        self._info("Pressing at (%s, %s)." % (coordinate_X, coordinate_Y))
-        driver = self._current_application()
-        action = TouchAction(driver)
-        action.press(x=coordinate_X, y=coordinate_Y).release().perform()
 
     def drag_and_drop(self, locator: str, target: str):
         """Drags the element identified by ``locator`` into the ``target`` element.
@@ -214,7 +294,7 @@ class _TouchKeywords(KeywordGroup):
         - ``origin`` - the element to drag
         - ``destination`` - the element to drag to
 
-        Usage:
+        Example:
         | `Drag And Drop` | id=div#element | id=div.target |
         """
         element = self._element_find(locator, True, True)
@@ -231,7 +311,7 @@ class _TouchKeywords(KeywordGroup):
         - ``end_x``   - x-coordinate at which to stop
         - ``end_y``   - y-coordinate at which to stop
 
-        Usage:
+        Example:
         | Flick | 100 | 100 | 100 | 400 | # Flicks the screen up. |
         """
         driver = self._current_application()